+++ CHANGELOG.md
@@ -3,4 +3,4 @@
 ### Fixed:
 - 
 ### Changed:
-- Updated to mc 1.20.1
+- Changed config from json5 to json